more refactoring to simplify dir structure
