+++ frontend/hooks/toast-if-api-not-connected.ts
