+++ frontend/hooks/use-global-store.ts
