+++ package.json
@@ -10,16 +10,16 @@
     "bootstrap:ci": "bun i --frozen-lockfile",
     "start": "bun cli.ts",
     "dev": "bun run build:dev-server && concurrently 'bun run dev:frontend' 'bun run dev:test-project'",
-    "dev:frontend": "vite dev --config dev-server-frontend/vite.config.ts",
+    "dev:frontend": "vite dev --config frontend/vite.config.ts",
     "dev:test-project": "bun --hot cli.ts dev --cwd ./example-project",
     "start:dev-server": "bun build:dev-server && bun cli.ts dev -y --cwd ./example-project",
-    "build:dev-server": "bun run build:dev-server:api && bun run build:dev-server:frontend",
-    "build:dev-server:api": "cd api && winterspec bundle-routes -i ./routes -o ./static-routes.ts",
-    "build:dev-server:frontend": "vite build --config dev-server-frontend/vite.config.ts && make-vfs --content-format string --dir dev-server-frontend/dist --outfile ./dev-server-frontend/dist/bundle.ts",
+    "build:dev-server": "bun run build:api && bun run build:frontend",
+    "build:api": "cd api && winterspec bundle-routes -i ./routes -o ./static-routes.ts",
+    "build:frontend": "vite build --config frontend/vite.config.ts && make-vfs --content-format string --dir frontend/dist --outfile ./frontend/dist/bundle.ts",
     "build:cli": "bun build-cli.ts",
     "build": "bun build:dev-server && npm run build:cli",
     "test:init": "bun cli.ts init --dir ./tmp/test --name test",
-    "update-deps": "bun add @tscircuit/builder@latest @tscircuit/react-fiber@latest && cd dev-server-frontend && bun run update-deps && cd ../example-project && bun run update-deps",
+    "update-deps": "bun add @tscircuit/builder@latest @tscircuit/react-fiber@latest @tscircuit/schematic-viewer@latest @tscircuit/pcb-viewer@latest @tscircuit/table-viewer@latest @tscircuit/3d-viewer@latest",
     "format": "biome format . --write"
   },
   "bin": {
@@ -113,6 +113,7 @@
     "@types/node": "^20.10.6",
     "@types/prompts": "^2.4.9",
     "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.3.0",
     "@types/semver": "^7.5.8",
     "@vitejs/plugin-react": "^4.3.1",
     "autoprefixer": "^10.4.20",
